requirements.txt
evdev==1.3.0
keyboard==0.13.5
Pillow==8.0.1
pynput==1.7.1
python-xlib==0.28
six==1.15.0
